+++ back/app/src/main/java/com/sloway/app/payment/stats/dto/response/MonthlyTrendResDto.java
@@ -8,9 +8,9 @@ import lombok.Getter;
 public class MonthlyTrendResDto {
 
     private String yearMonth;
-    private Integer totalAmt;
+    private Long totalAmt;   // 월 매출/건수 — int 오버플로우 방지로 Long
 
-    public static MonthlyTrendResDto of(String yearMonth, Integer totalAmt) {
+    public static MonthlyTrendResDto of(String yearMonth, Long totalAmt) {
         return MonthlyTrendResDto.builder()
                 .yearMonth(yearMonth)
                 .totalAmt(totalAmt)
